// src/app/models/itickets-info.ts
export interface ITicketsInfo {
    id: number;
    name: string;
    ticketNumber: string;
    price: number;
    cinemas: string;
    peopleCount: number;
    purchaseCount: string;
    ticket: string;
    phoneNumber:string;
    cinemasLocation:string;
    image?:string;
    ticketDetails:string;
}